requirements.txt
pytest
Jinja2
pyyaml
Flask
flask-restful
